+++ extension.js
@@ -12,7 +12,13 @@ let defaultDir = 'src/components',
 	addConstructor = false,
 	componentType = 'class',
 	openAfterCreate = true,
-	language = 'en';
+	language = 'en',
+	componentNameStyle = {
+		styleInExplorer: 'camel',
+		bigFirstLetterInExplorer: false,
+		styleInCode: 'camel',
+		bigFirstLetterInCode: true 
+	}
 
 	const inputComponentNamePlaceholder = {
 		en: 'Enter the name of the component in Latin, separating the words with a space',
@@ -58,15 +64,79 @@ function readConfig() {
 
 function jsonParse(json) {
 	const config = JSON.parse(json);
+	console.log(config);
 	defaultDir = config.defaultDir || 'src/components';
 	addCSS = (config.addCSS != undefined) ? config.addCSS : true;
 	addConstructor = (config.addConstructor != undefined) ? config.addConstructor : false;
 	componentType = (config.componentType == 'function') ? 'function' : 'class';
-	openAfterCreate = (config.openAfterCreate != undefined) ? config.openAfterCreate : true,
+	openAfterCreate = (config.openAfterCreate != undefined) ? config.openAfterCreate : true;
 	language = (config.addCSS == undefined || config.language != 'ru') ? 'en' : 'ru';
+	componentNameStyle.styleInExplorer = (
+		config.componentNameStyle.styleInExplorer == undefined &&
+		config.componentNameStyle.styleInExplorer != 'snake' &&
+		config.componentNameStyle.styleInExplorer != 'kebab'
+	) ? 'camel' : config.componentNameStyle.styleInExplorer;
+	console.log('config.componentNameStyle.styleInExplorer=' + config.componentNameStyle.styleInExplorer);
+	console.log('componentNameStyle.styleInExplorer=' + componentNameStyle.styleInExplorer);
+	componentNameStyle.styleInCode = (
+		config.componentNameStyle.styleInCode == undefined &&
+		config.componentNameStyle.styleInCode != 'snake' &&
+		config.componentNameStyle.istyleInCode != 'kebab'
+	) ? 'camel' : config.componentNameStyle.styleInCode;
+	componentNameStyle.bigFirstLetterInExplorer = (config.componentNameStyle.bigFirstLetterInExplorer != undefined) ? config.componentNameStyle.bigFirstLetterInExplorer : false;
+	componentNameStyle.bigFirstLetterInCode = (config.componentNameStyle.bigFirstLetterInCode != undefined) ? config.componentNameStyle.bigFirstLetterInCode : true;
+}
+
+function textStyle(text, style = 'camel', firstLetterBig = true) {
+	const splitedText = text.toLowerCase().split(' ');
+	let newText = '';
+	style = style.toLowerCase();
+	if (style == 'camel') {
+		splitedText.forEach((word, i) => {
+			if (i == 0 && !firstLetterBig) {
+				newText += word;
+			} else {
+				newText += word[0].toUpperCase() + word.slice(1);
+			}
+		});
+	} else if (style == 'snake') {
+		splitedText.forEach((word, i) => {
+			if (i == 0) {
+				if (firstLetterBig) {
+					newText = word[0].toUpperCase() + word.slice(1);
+				} else {
+					newText = word;
+				}
+			} else {
+				newText += '_' + word;
+			}
+		});
+	} else if (style == 'kebab') {
+		splitedText.forEach((word, i) => {
+			if (i == 0) {
+				if (firstLetterBig) {
+					newText = word[0].toUpperCase() + word.slice(1);
+				} else {
+					newText = word;
+				}
+			} else {
+				newText += '-' + word;
+			}
+		});
+	} else {
+		splitedText.forEach((word, i) => {
+			if (i == 0 && firstLetterBig) {
+				newText += word[0].toUpperCase() + word.slice(1);
+			} else {
+				newText += word;
+			}
+		});
+	}
+	return newText;
 }
 
 function createComponent(dir) {
+	readConfig();
 	vscode.window.showInputBox({
 		placeHolder: inputComponentNamePlaceholder[language]
 	}).then((input) => {
@@ -74,20 +144,20 @@ function createComponent(dir) {
 		if (!firstSymbolIsLetter(input)) {
 			vscode.window.showErrorMessage(wrongComponentNameError[language]);
 		} else {
-			readConfig();
-			const splitedInput = input.toLowerCase().split(' ');
-			let componentName = '';
-			splitedInput.forEach(word => {
-				componentName += word[0].toUpperCase() + word.slice(1);
-			});
+			const componentName = textStyle(input, componentNameStyle.styleInCode, componentNameStyle.bigFirstLetterInCode);
+			// const splitedInput = input.toLowerCase().split(' ');
+			// let componentName = '';
+			// splitedInput.forEach(word => {
+			// 	componentName += word[0].toUpperCase() + word.slice(1);
+			// });
+			const componentDirName = textStyle(input, componentNameStyle.styleInExplorer, componentNameStyle.bigFirstLetterInExplorer);
 
-			const componentDirName = componentName[0].toLowerCase() + componentName.slice(1);
+			//const componentDirName = componentName[0].toLowerCase() + componentName.slice(1);
 
 			const CSSCommand = addCSS ? `import './${componentDirName}.css';\n\n` : '';
 			const constructorCommand = addConstructor ? `	constructor(props) {\n		super(props);\n	}\n\n` : '';
 			
-			const contentIndex = 
-			`import ${componentName} from './${componentDirName}';\nexport default ${componentName};`;
+			const contentIndex = `import ${componentName} from './${componentDirName}';\nexport default ${componentName};`;
 			
 			const contentComponent = {
 				class: `import React, {Component} from 'react';\n\n${CSSCommand}export default class ${componentName} extends Component {\n${constructorCommand}	render() {\n		return (\n			<>\n\n			</>\n		)\n	}\n}`,
@@ -150,7 +220,7 @@ function activate(context) {
 	}
 
 	const createDefaultConfig = () => {
-		const contentConfig = `{\n	"defaultDir": "src/components",\n	"componentType": "class",\n	"openAfterCreate": true,\n	"addCSS": true,\n	"addConstructor": false,\n	"language": "en"\n}`;
+		const contentConfig = `{\n	"defaultDir": "src/components",\n	"componentType": "class",\n	"openAfterCreate": true,\n	"addCSS": true,\n	"addConstructor": false,\n	"language": "en",\n	"componentNameStyle": {\n		"styleInExplorer": "camel",\n		"bigFirstLetterInExplorer": false,\n		"styleInCode": "camel",\n		"bigFirstLetterInCode": true\n	}\n}`;
 
 		fs.mkdirSync(path.join(projectDir, `.vscode`), { recursive: true });